Forage Virtual Internship
This project aims to analyze the hotel revenue and customer experience for a hotel chain over the years 2018, 2019, and 2020. The project use SQL queries to analyze the data and Power BI to identify key trends. The project also make recommendations for how the hotel chain can improve its revenue and customer experience.